Final Version
uses the ineffective debouncing strategy

+++ masterNodeAlpha/functionFrameTracker.ino
@@ -1,7 +1,7 @@
 int frameTracker()
 {
   // should always round down to current frame
-  int currentFrame = (millis() % masterWrapTime) / frameDuration;
+  int currentFrame = (millis() % ((unsigned long) masterWrapTime)) / ((unsigned long) frameDuration);
   
   //  has the frame changed since last check
   if (currentFrame != lastAccessedFrame)
